--- src/components/Visualization/habitat_risk_assessment/actions/index.js
@@ -1,14 +1,0 @@
-export const getCsvUrl = csvUrl => ({
-  type: 'CSV_UPLOADED',
-  payload: csvUrl
-});
-
-export const getVectorsOnMap = vectorsOnMap => ({
-  type: 'VECTORS_UPDATED',
-  payload: vectorsOnMap
-});
-
-export const getFileSuffix = fileSuffix => ({
-  type: 'SUFFIX_OBTAINED',
-  payload: fileSuffix
-});
