+++ src/components/VizTab/Visualization/habitat_risk_assessment/Map/Control.js
