+++ src/components/VizTab/Visualization/habitat_risk_assessment/Plot/Legend.js
