+++ src/components/VizTab/Visualization/habitat_risk_assessment/reducers/index.js
