+++ src/components/VizTab/VizErrorBoundary.jsx
