+++ src/index.html
@@ -22,11 +22,11 @@
   />
   <link
     rel="stylesheet"
-    href="./components/Visualization/habitat_risk_assessment/Map/style.css"
+    href="./components/VizTab/Visualization/habitat_risk_assessment/Map/style.css"
   />
   <link
     rel="stylesheet"
-    href="./components/Visualization/habitat_risk_assessment/Plot/style.css"
+    href="./components/VizTab/Visualization/habitat_risk_assessment/Plot/style.css"
   />
 </head>
